requirements.txt
flask
openai
num2words
matplotlib
plotly
scipy
scikit-learn
pandas
tiktoken
PyMuPDF
pillow
pytesseract
markdown
psycopg2-binary
python-dotenv
gunicorn